tests/wpt/mozilla/tests/bluetooth/requestDevice/canonicalizeFilter/no-filters-member.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.empty);
    return promise_rejects_js(
        t,
        TypeError,
        window.navigator.bluetooth.requestDevice([{}, {optionalServices: ['wrong_service']}]));
}, 'RequestDeviceOptions requires a |filters| member.');
</script>
